Using moonbeam types entirely
