+++ substrate/primitives/core/Cargo.toml
@@ -52,7 +52,7 @@ blake2 = { version = "0.10.4", default-features = false, optional = true }
 libsecp256k1 = { version = "0.7", default-features = false, features = ["static-context"], optional = true }
 schnorrkel = { version = "0.9.1", features = ["preaudit_deprecated", "u64_backend"], default-features = false }
 merlin = { version = "2.0", default-features = false }
-secp256k1 = { version = "0.24.0", default-features = false, features = ["recovery", "alloc"]}
+secp256k1 = { version = "0.24.0", default-features = false, features = ["recovery", "alloc"], optional = true }
 sp-core-hashing = { path = "hashing", default-features = false, optional = true }
 sp-runtime-interface = { path = "../runtime-interface", default-features = false}
 
@@ -147,6 +147,7 @@ full_crypto = [
 	"blake2",
 	"ed25519-zebra",
 	"libsecp256k1",
+	"secp256k1",
 	"sp-core-hashing",
 	"sp-runtime-interface/disable_target_static_assertions",
 ]
