+++ substrate/primitives/core/src/crypto.rs
@@ -40,7 +40,6 @@ use sp_std::{hash::Hash, str, vec::Vec};
 pub use ss58_registry::{from_known_address_format, Ss58AddressFormat, Ss58AddressFormatRegistry};
 /// Trait to zeroize a memory buffer.
 pub use zeroize::Zeroize;
-use crate::hash::H160;
 
 /// The root phrase for our publicly known keys.
 pub const DEV_PHRASE: &str =
@@ -628,130 +627,6 @@ impl sp_std::str::FromStr for AccountId32 {
 		} else {
 			Self::from_ss58check(s).map_err(|_| "invalid ss58 address.")
 		}
-	}
-}
-
-#[derive(
-	Eq, PartialEq, Copy, Clone, Encode, Decode, TypeInfo, MaxEncodedLen, Default, PartialOrd, Ord,
-)]
-pub struct AccountId20(pub [u8; 20]);
-impl AccountId20 {
-	/// Create a new instance from its raw inner byte value.
-	///
-	/// Equivalent to this types `From<[u8; 32]>` implementation. For the lack of const
-	/// support in traits we have this constructor.
-	pub const fn new(inner: [u8; 20]) -> Self {
-		Self(inner)
-	}
-}
-impl_serde::impl_fixed_hash_serde!(AccountId20, 20);
-
-impl UncheckedFrom<crate::hash::H160> for AccountId20 {
-	fn unchecked_from(h: crate::hash::H160) -> Self {
-		AccountId20(h.into())
-	}
-}
-
-impl ByteArray for AccountId20 {
-	const LEN: usize = 20;
-}
-
-#[cfg(feature = "std")]
-impl std::fmt::Display for AccountId20 {
-	//TODO This is a pretty quck-n-dirty implementation. Perhaps we should add
-	// checksum casing here? I bet there is a crate for that.
-	// Maybe this one https://github.com/miguelmota/rust-eth-checksum
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		write!(f, "{:?}", self.0)
-	}
-}
-
-impl core::fmt::Debug for AccountId20 {
-	fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-		write!(f, "{:?}", H160(self.0))
-	}
-}
-
-impl AsRef<[u8]> for AccountId20 {
-	fn as_ref(&self) -> &[u8] {
-		&self.0[..]
-	}
-}
-
-impl AsMut<[u8]> for AccountId20 {
-	fn as_mut(&mut self) -> &mut [u8] {
-		&mut self.0[..]
-	}
-}
-
-impl AsRef<[u8; 20]> for AccountId20 {
-	fn as_ref(&self) -> &[u8; 20] {
-		&self.0
-	}
-}
-
-impl AsMut<[u8; 20]> for AccountId20 {
-	fn as_mut(&mut self) -> &mut [u8; 20] {
-		&mut self.0
-	}
-}
-
-impl<'a> TryFrom<&'a [u8]> for AccountId20 {
-	type Error = ();
-	fn try_from(x: &'a [u8]) -> Result<AccountId20, ()> {
-		if x.len() == 20 {
-			let mut data = [0; 20];
-			data.copy_from_slice(x);
-			Ok(AccountId20(data))
-		} else {
-			Err(())
-		}
-	}
-}
-
-impl From<[u8; 20]> for AccountId20 {
-	fn from(bytes: [u8; 20]) -> Self {
-		Self(bytes)
-	}
-}
-
-impl From<AccountId20> for [u8; 20] {
-	fn from(value: AccountId20) -> Self {
-		value.0
-	}
-}
-
-// NOTE: the implementation is lossy, and is intended to be used
-// only to convert from Polkadot accounts to AccountId20.
-// See https://github.com/moonbeam-foundation/moonbeam/pull/2315#discussion_r1205830577
-// DO NOT USE IT FOR ANYTHING ELSE.
-impl From<[u8; 32]> for AccountId20 {
-	fn from(bytes: [u8; 32]) -> Self {
-		let mut buffer = [0u8; 20];
-		buffer.copy_from_slice(&bytes[..20]);
-		Self(buffer)
-	}
-}
-
-impl From<H160> for AccountId20 {
-	fn from(h160: H160) -> Self {
-		Self(h160.0)
-	}
-}
-
-impl From<AccountId20> for H160 {
-	fn from(value: AccountId20) -> Self {
-		H160(value.0)
-	}
-}
-
-#[cfg(feature = "std")]
-impl std::str::FromStr for AccountId20 {
-	type Err = &'static str;
-	fn from_str(input: &str) -> Result<Self, Self::Err> {
-		H160::from_str(input)
-			.map(Into::into)
-			.map_err(|_| "invalid hex address.")
 	}
 }
